+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - `queryChanged` behavior has changed slightly: If an entity's storage was changed multiple times since your system last observed it, the `old` field in the `ChangeRecord` will be the last value your system observed the entity as having for that component, rather than what it was most recently changed from.
 - World and Loop types are now exported (#9)
+- Matter now uses both `__iter` and `__call` for iteration over `QueryResult`.
 
 ### Fixed
 - System error stack traces are now displayed properly (#12)